Error Handling + Styles
Added a page for 400 and 404 errors, updated search routes so that if a title is not provided it gives a 400 error. Improved Styling of the temp search. Added the footer to all pages with working back to top. Also there's now a script for starting the development enviornment (Windows only)